pyproject.toml
[project]
name = "gemini"
version = "0.1.0"
description = ""
authors = [
    { name = "Chris", email = "chris@example.com" }
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "websockets (>=15.0.1,<16.0.0)",
    "pymongo (>=4.14.0,<5.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
